#27: Fixes
