+++ lib/common/ui/macos_ui_hacks.dart
@@ -1,0 +1,226 @@
+import 'package:flutter/cupertino.dart';
+import 'package:macos_ui/macos_ui.dart';
+
+/// –––––––––––––––––––––––––––––––––––––––––––––––––––––
+/// ––– HERE AND BELOW JUST COPY-PASTED FROM MACOS_UI –––
+/// –––––––––––– BECAUSE IT IS PRIVATE THERE ––––––––––––
+/// –––––––––––––––––––––––––––––––––––––––––––––––––––––
+
+class MacosSidebarItem extends StatelessWidget {
+  /// Builds a [MacosSidebarItem].
+  // ignore: use_super_parameters
+  const MacosSidebarItem({
+    required this.item,
+    required this.onClick,
+    required this.selected,
+    Key? key,
+  }) : super(key: key);
+
+  /// The widget to lay out first.
+  ///
+  /// Typically an [Icon]
+  final SidebarItem item;
+
+  /// Whether the item is selected or not
+  final bool selected;
+
+  /// A function to perform when the widget is clicked or tapped.
+  ///
+  /// Typically a [Navigator] call
+  final VoidCallback? onClick;
+
+  void _handleActionTap() => onClick?.call();
+
+  Map<Type, Action<Intent>> get _actionMap => <Type, Action<Intent>>{
+        ActivateIntent: CallbackAction<ActivateIntent>(
+          onInvoke: (intent) => _handleActionTap(),
+        ),
+        ButtonActivateIntent: CallbackAction<ButtonActivateIntent>(
+          onInvoke: (intent) => _handleActionTap(),
+        ),
+      };
+
+  bool get hasLeading => item.leading != null;
+
+  bool get hasTrailing => item.trailing != null;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = MacosTheme.of(context);
+
+    final selectedColor = MacosDynamicColor.resolve(
+      MacosColorProvider.getActiveColor(
+        accentColor: theme.accentColor ?? AccentColor.blue,
+        isDarkModeEnabled: theme.brightness == Brightness.dark,
+        isWindowMain: true,
+      ),
+      context,
+    );
+    final unselectedColor = MacosDynamicColor.resolve(
+      item.unselectedColor ?? MacosColors.transparent,
+      context,
+    );
+
+    final double spacing = 10.0 + theme.visualDensity.horizontal;
+    const itemSize = SidebarItemSize.medium;
+    TextStyle? labelStyle;
+    switch (itemSize) {
+      case SidebarItemSize.small:
+        labelStyle = theme.typography.subheadline;
+        break;
+      case SidebarItemSize.medium:
+        labelStyle = theme.typography.body;
+        break;
+      case SidebarItemSize.large:
+        labelStyle = theme.typography.title3;
+        break;
+    }
+
+    return Semantics(
+      label: item.semanticLabel,
+      button: true,
+      focusable: true,
+      focused: item.focusNode?.hasFocus,
+      enabled: onClick != null,
+      selected: selected,
+      child: GestureDetector(
+        onTap: onClick,
+        child: FocusableActionDetector(
+          focusNode: item.focusNode,
+          descendantsAreFocusable: false,
+          enabled: onClick != null,
+          //mouseCursor: SystemMouseCursors.basic,
+          actions: _actionMap,
+          child: Container(
+            width: 134.0 + theme.visualDensity.horizontal,
+            height: itemSize.height + theme.visualDensity.vertical,
+            decoration: ShapeDecoration(
+              color: selected ? selectedColor : unselectedColor,
+              shape: item.shape ??
+                  const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(5)),
+                  ),
+            ),
+            padding: EdgeInsets.symmetric(
+              vertical: 7 + theme.visualDensity.horizontal,
+              horizontal: spacing,
+            ),
+            child: Row(
+              children: [
+                if (hasLeading)
+                  Padding(
+                    padding: EdgeInsets.only(right: spacing),
+                    child: MacosIconTheme.merge(
+                      data: MacosIconThemeData(
+                        color:
+                            selected ? MacosColors.white : theme.primaryColor,
+                        size: itemSize.iconSize,
+                      ),
+                      child: item.leading!,
+                    ),
+                  ),
+                Expanded(
+                  child: DefaultTextStyle(
+                    style: labelStyle.copyWith(
+                      color: selected ? textLuminance(selectedColor) : null,
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                    child: item.label,
+                  ),
+                ),
+                if (hasTrailing) ...[
+                  const Spacer(),
+                  DefaultTextStyle(
+                    style: labelStyle.copyWith(
+                      color: selected ? textLuminance(selectedColor) : null,
+                    ),
+                    child: item.trailing!,
+                  ),
+                ],
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+Color textLuminance(Color backgroundColor) {
+  return backgroundColor.computeLuminance() >= 0.5
+      ? CupertinoColors.black
+      : CupertinoColors.white;
+}
+
+class MacosColorProvider {
+  MacosColorProvider._();
+
+  /// Returns the active color based on the provided parameters.
+  static Color getActiveColor({
+    required AccentColor accentColor,
+    required bool isDarkModeEnabled,
+    required bool isWindowMain,
+  }) {
+    if (isDarkModeEnabled) {
+      if (!isWindowMain) {
+        return const MacosColor.fromRGBO(76, 78, 65, 1);
+      }
+
+      switch (accentColor) {
+        case AccentColor.blue:
+          return const MacosColor.fromRGBO(22, 105, 229, 0.749);
+
+        case AccentColor.purple:
+          return const MacosColor.fromRGBO(204, 45, 202, 0.749);
+
+        case AccentColor.pink:
+          return const MacosColor.fromRGBO(229, 74, 145, 0.749);
+
+        case AccentColor.red:
+          return const MacosColor.fromRGBO(238, 64, 68, 0.749);
+
+        case AccentColor.orange:
+          return const MacosColor.fromRGBO(244, 114, 0, 0.749);
+
+        case AccentColor.yellow:
+          return const MacosColor.fromRGBO(233, 176, 0, 0.749);
+
+        case AccentColor.green:
+          return const MacosColor.fromRGBO(76, 177, 45, 0.749);
+
+        case AccentColor.graphite:
+          return const MacosColor.fromRGBO(129, 129, 122, 0.824);
+      }
+    }
+
+    if (!isWindowMain) {
+      return const MacosColor.fromRGBO(180, 180, 180, 1);
+    }
+
+    switch (accentColor) {
+      case AccentColor.blue:
+        return const MacosColor.fromRGBO(9, 129, 255, 0.749);
+
+      case AccentColor.purple:
+        return const MacosColor.fromRGBO(162, 28, 165, 0.749);
+
+      case AccentColor.pink:
+        return const MacosColor.fromRGBO(234, 81, 152, 0.749);
+
+      case AccentColor.red:
+        return const MacosColor.fromRGBO(220, 32, 40, 0.749);
+
+      case AccentColor.orange:
+        return const MacosColor.fromRGBO(245, 113, 0, 0.749);
+
+      case AccentColor.yellow:
+        return const MacosColor.fromRGBO(240, 180, 2, 0.749);
+
+      case AccentColor.green:
+        return const MacosColor.fromRGBO(66, 174, 33, 0.749);
+
+      case AccentColor.graphite:
+        return const MacosColor.fromRGBO(174, 174, 167, 0.847);
+    }
+  }
+}
