+++ lib/features/tools/message/tools_message.dart
@@ -1,6 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../../tool/tool.dart';
 import '../state/tools_state.dart';
 
 part 'tools_message.freezed.dart';
